move decoding to a tools class,
EID translating will be handled by subsystems
